feat: making db optional

+++ gridoai_ml/app.py
@@ -14,6 +14,8 @@ database = get_database(setup_data, model.dim)
 
 @app.post("/write")
 async def write(document: Document):
+    if database is None:
+        return {"message": "Running without database"}
     print(f"Received document: {document}")
     vec = model.calc(document.content)
     if vec is not None:
@@ -26,16 +28,28 @@ async def write(document: Document):
 
 @app.get("/delete")
 async def delete(uid: UUID):
+    if database is None:
+        return {"message": "Running without database"}
     print(f"Received uid: {uid}")
     database.delete_doc(uid)
 
 
 @app.get("/neardocs")
 async def neardocs(text: str):
+    if database is None:
+        return {"message": "Running without database"}
     print(f"Received text: {text}")
     vec = model.calc(text)
     if vec is not None:
         docs = database.get_near_vecs(vec.tolist(), 10)
         return {"message": docs}
+    else:
+        return {"message": "error"}
+
+@app.get("/embed")
+async def embed(text: str):
+    vec = model.calc(text)
+    if vec is not None:
+        return {"message": vec.tolist()}
     else:
         return {"message": "error"}
